Crate cu29_export

Structs

LogReaderCli

This is a generator for a main function to build a log extractor.

PyCuLogEntry

This is a python wrapper for CuLogEntries.

PyLogIterator

Enums

Command

ExportFormat

Functions

copperlists_dump

Extracts the copper lists from a binary representation. P is the Payload determined by the configuration of the application.

run_cli

This is a generator for a main function to build a log extractor. It depends on the specific type of the CopperList payload that is determined at compile time from the configuration.

struct_log_iterator_bare

Creates an iterator of CuLogEntries from a bare binary structured log file (ie. not within a unified log). This is mainly used for using the structured logging out of the Copper framework. it returns a tuple with the iterator of log entries and the list of interned strings.

struct_log_iterator_unified

Creates an iterator of CuLogEntries from a unified log file. This function allows you to easily use python to datamind Copper’s structured text logs. it returns a tuple with the iterator of log entries and the list of interned strings.

textlog_dump

Full dump of the copper structured log from its binary representation. This rebuilds a textual log. src: the source of the log data index: the path to the index file (containing the interned strings constructed at build time)