cu29_export
Function run_cli

pub fn run_cli<P>() -> CuResult<()>
where
    P: CopperListTuple,

This is a generator for a main function to build a log extractor. It depends on the specific type of the CopperList payload that is determined at compile time from the configuration.