Function cu29_export::struct_log_iterator_unified

pub fn struct_log_iterator_unified(
    unified_src_path: &str,
    index_path: &str,
) -> PyResult<(PyLogIterator, Vec<String>)>

Creates an iterator of CuLogEntries from a unified log file. This function allows you to easily use python to datamind Copper’s structured text logs. it returns a tuple with the iterator of log entries and the list of interned strings.