Crate cu29_export

Structs

LogReaderCli

This is a generator for a main function to build a log extractor.

Enums

Command

ExportFormat

Functions

copperlists_reader

Extracts the copper lists from a binary representation. P is the Payload determined by the configuration of the application.

keyframes_reader

Extracts the keyframes from the log.

run_cli

This is a generator for a main function to build a log extractor. It depends on the specific type of the CopperList payload that is determined at compile time from the configuration.

structlog_reader

textlog_dump

Full dump of the copper structured log from its binary representation. This rebuilds a textual log. src: the source of the log data index: the path to the index file (containing the interned strings constructed at build time)